refactor: subscription 폴더 상위 루트로 이동
